Moved full project code into Main.java
