+++ Java/src/Main.java
@@ -1,6 +1,44 @@
-import algorithms.BubbleSort;
-import algorithms.QuickSort;
-import algorithms.SortContext;
+// Schnittstelle für die Sortierstrategie
+interface SortStrategy {
+    void sort(int[] array);
+}
+
+// Konkrete Strategie: Bubble Sort
+class BubbleSort implements SortStrategy {
+    @Override
+    public void sort(int[] array) {
+        // Implementierung des Bubble Sort Algorithmus
+        // ...
+        System.out.println("Bubble Sort angewendet.");
+    }
+}
+
+// Konkrete Strategie: Quick Sort
+class QuickSort implements SortStrategy {
+    @Override
+    public void sort(int[] array) {
+        // Implementierung des Quick Sort Algorithmus
+        // ...
+        System.out.println("Quick Sort angewendet.");
+    }
+}
+
+// Kontext, der die Sortierstrategie verwendet
+class SortContext {
+    private SortStrategy strategy;
+
+    public SortContext(SortStrategy strategy) {
+        this.strategy = strategy;
+    }
+
+    public void setStrategy(SortStrategy strategy) {
+        this.strategy = strategy;
+    }
+
+    public void executeSort(int[] array) {
+        strategy.sort(array);
+    }
+}
 
 public class Main {
     public static void main(String[] args) {
