--- Java/src/algorithms/BubbleSort.java
@@ -1,13 +1,0 @@
-package algorithms;
-
-import strategy.SortStrategy;
-
-// Konkrete Strategie: Bubble Sort
-public class BubbleSort implements SortStrategy {
-    @Override
-    public void sort(int[] array) {
-        // Implementierung des Bubble Sort Algorithmus
-        // ...
-        System.out.println("Bubble Sort angewendet.");
-    }
-}
