--- Java/src/algorithms/QuickSort.java
@@ -1,13 +1,0 @@
-package algorithms;
-
-import strategy.SortStrategy;
-
-// Konkrete Strategie: Quick Sort
-public class QuickSort implements SortStrategy {
-    @Override
-    public void sort(int[] array) {
-        // Implementierung des Quick Sort Algorithmus
-        // ...
-        System.out.println("Quick Sort angewendet.");
-    }
-}
