--- Java/src/strategy/SortStrategy.java
@@ -1,6 +1,0 @@
-package strategy;
-
-// Schnittstelle für die Sortierstrategie
-public interface SortStrategy {
-    void sort(int[] array);
-}
